[c10d] Monitored barrier: option to collect all failed ranks
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55010

Follow up change to add a flag to provide an option for monitored barrier to collect all the failed ranks and then throw instead of just throwing on the first one. This is useful as now monitored barrier will be able to pick up on all hanging ranks instead of just one.

This is done by passing in a flag `wait_all_ranks=True`.
ghstack-source-id: 125326669

Differential Revision: [D27447787](https://our.internmc.facebook.com/intern/diff/D27447787/)