[TensorExpr] Nuke DepTracker and findAllNeededTensors.
DepTracker was used to automatically pull in dependent computations from
output ones. While it seems quite convenient, it's led to several
architectural issues, which are fixed in this stack.

DepTracker worked on Tensors, which is a pair of Buf and Stmt. However,
Stmt could become stale and there was no way to reliably update the
corresponding tensor. We're now using Bufs and Stmts directly and moving
away from using Tensors to avoid these problems.

Removing DepTracker allowed to unify Loads and FunctionCalls, which
essentially were duplicates of each other.

ghstack-source-id: 8bdf6318513986ddbd804d6434ffe03782a81ac3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54997